Pass the forward test

+++ csrc/dwconv.cu
@@ -15,15 +15,15 @@ __device__ void block_reduce(Dtype* data) {
 template <typename Dtype>
 __global__ void dwconv_forward_backward_kernel(
     Dtype* out, const Dtype* __restrict__ data,
-    const Dtype* __restrict__ weight, const int* __restrict__ neigh,
+    const Dtype* __restrict__ weight, const int64_t* __restrict__ neigh,
     const int64_t height, const int64_t channel, const int64_t kngh,
     const int64_t nthreads) {
   CUDA_KERNEL_LOOP(i, nthreads) {
-    int h = i / channel;
-    int c = i % channel;
+    int64_t h = i / channel;
+    int64_t c = i % channel;
     Dtype value = Dtype(0);
     for (int k = 0; k < kngh; ++k) {
-      int ni = neigh[h * kngh + k];
+      int64_t ni = neigh[h * kngh + k];
       if (ni >= 0) {
         value += weight[k * channel + c] * data[ni * channel + c];
       }
@@ -35,15 +35,15 @@ __global__ void dwconv_forward_backward_kernel(
 template <typename Dtype>
 __global__ void dwconv_weight_backward_kernel(
     Dtype* out, const Dtype* __restrict__ grad, const Dtype* __restrict__ data,
-    const int* __restrict__ ineigh, const int64_t height, const int64_t channel,
+    const int64_t* __restrict__ ineigh, const int64_t height, const int64_t channel,
     const int64_t kngh, const int64_t nthreads) {
   __shared__ Dtype weights[kCudaThreadsNum];
   CUDA_KERNEL_LOOP(i, nthreads) {
-    int h = i / channel;
-    int c = i % channel;
+    int64_t h = i / channel;
+    int64_t c = i % channel;
     for (int k = 0; k < kngh; ++k) {
-      int ni = ineigh[h * kngh + k];
-      int tid = threadIdx.x;
+      int64_t ni = ineigh[h * kngh + k];
+      int64_t tid = threadIdx.x;
       weights[tid] = ni >= 0 ? data[ni * channel + c] * grad[i] : Dtype(0);
       __syncthreads();
       block_reduce(weights);
@@ -56,12 +56,12 @@ __global__ void dwconv_weight_backward_kernel(
 }
 
 __global__ void inverse_neigh_kernel(
-    int* ineigh, const int* __restrict__ neigh, const int64_t height, 
+    int64_t* ineigh, const int64_t* __restrict__ neigh, const int64_t height, 
     const int64_t kngh, const int64_t nthreads) {
   CUDA_KERNEL_LOOP(i, nthreads) {
-    int h = i / kngh;
-    int c = i % kngh;
-    int j = neigh[i];
+    int64_t h = i / kngh;
+    int64_t c = i % kngh;
+    int64_t j = neigh[i];
     if (j >= 0) {
       ineigh[j * kngh + c] = h;
     }
@@ -76,7 +76,7 @@ Tensor inverse_neigh(Tensor neigh) {
   auto stream = at::cuda::getCurrentCUDAStream();
   inverse_neigh_kernel
       <<<CudaGetBlocks(nthreads), kCudaThreadsNum, 0, stream>>>(
-      ineigh.data_ptr<int>(), neigh.data_ptr<int>(), height, kngh, nthreads);
+      ineigh.data_ptr<int64_t>(), neigh.data_ptr<int64_t>(), height, kngh, nthreads);
   return ineigh;
 }
 
@@ -91,7 +91,7 @@ Tensor dwconv_forward_backward(Tensor data, Tensor weight, Tensor neigh) {
   dwconv_forward_backward_kernel<float>
       <<<CudaGetBlocks(nthreads), kCudaThreadsNum, 0, stream>>>(
       out.data_ptr<float>(), data.data_ptr<float>(), weight.data_ptr<float>(),
-      neigh.data_ptr<int>(), height, channel, kngh, nthreads);
+      neigh.data_ptr<int64_t>(), height, channel, kngh, nthreads);
   CUDA_POST_KERNEL_CHECK;
   return out;
 }
@@ -106,7 +106,7 @@ Tensor dwconv_weight_backward(Tensor grad, Tensor data, Tensor neigh) {
   dwconv_weight_backward_kernel<float>
       <<<CudaGetBlocks(nthreads), kCudaThreadsNum, 0, stream>>>(
       out.data_ptr<float>(), grad.data_ptr<float>(), data.data_ptr<float>(),
-      neigh.data_ptr<int>(), height, channel, kngh, nthreads);
+      neigh.data_ptr<int64_t>(), height, channel, kngh, nthreads);
   CUDA_POST_KERNEL_CHECK;
   return out.sum(0);
 }
